+++ includes/requirements.php
@@ -16,24 +16,40 @@
 	Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-require_once CBQE_DIR_LIB . 'aihrus-framework/requirements.php';
+require_once ABSPATH . 'wp-admin/includes/plugin.php';
+
+if ( ! function_exists( 'aihr_notice_version' ) ) {
+	function aihr_notice_version( $required_base, $required_name, $required_slug, $required_version, $item_name ) {
+		$is_active = is_plugin_active( $required_base );
+		if ( $is_active )
+			$link = sprintf( __( '<a href="%1$s">update to</a>', 'gc-testimonials-to-testimonials' ), self_admin_url( 'update-core.php' ) );
+		else {
+			$plugins = get_plugins();
+			if ( empty( $plugins[ $required_base ] ) ) {
+				$install = esc_url( wp_nonce_url( self_admin_url( 'update.php?action=install-plugin&plugin=' . $required_slug ), 'install-plugin_' . $required_slug ) );
+				$link    = sprintf( __( '<a href="%1$s">install</a>', 'gc-testimonials-to-testimonials' ), $install );
+			} else {
+				$activate = esc_url( wp_nonce_url( admin_url( 'plugins.php?action=activate&plugin=' . $required_base ), 'activate-plugin_' . $required_base ) );
+				$link     = sprintf( __( '<a href="%1$s">activate</a>', 'gc-testimonials-to-testimonials' ), $activate );
+			}
+		}
+
+		$text = sprintf( __( 'Plugin "%3$s" has been deactivated. Please %1$s "%4$s" version %2$s or newer before activating "%3$s".', 'gc-testimonials-to-testimonials' ), $link, $required_version, $item_name, $required_name );
+
+		$content  = '<div class="error"><p>';
+		$content .= $text;
+		$content .= '</p></div>';
+
+		echo $content;
+	}
+}
 
 
 function cbqe_ef_requirements_check() {
 	$valid_requirements = true;
-	if ( ! aihr_check_php( CBQE_EF_BASE, CBQE_EF_NAME ) ) {
+	if ( ! is_plugin_active( CBQE_EF_REQ_BASE ) && ! is_plugin_active( CBQE_EF_REQ_BASE_PREM ) ) {
 		$valid_requirements = false;
-	}
-
-	if ( ! aihr_check_wp( CBQE_EF_BASE, CBQE_EF_NAME ) ) {
-		$valid_requirements = false;
-	}
-
-	if ( ! is_plugin_active( CBQE_EF_REQ_BASE ) ) {
-		if ( ! is_plugin_active( CBQE_EF_REQ_BASE_PREM ) ) {
-			$valid_requirements = false;
-			add_action( 'admin_notices', 'cbqe_ef_notice_version' );
-		}
+		add_action( 'admin_notices', 'cbqe_ef_notice_version' );
 	}
 
 	if ( ! is_plugin_active( CBQE_EF_EXT_BASE ) ) {